Change speed to float

+++ db/migrate/20220314172122_add_speed_to_tracks.rb
@@ -1,5 +1,5 @@
 class AddSpeedToTracks < ActiveRecord::Migration[6.1]
   def change
-    add_column :tracks, :speed, :integer, array:true, default: []
+    add_column :tracks, :speed, :float, array:true, default: []
   end
 end
